+++ src/main/java/com/aws/greengrass/mqtt/bridge/clients/MQTTClientException.java
@@ -1,0 +1,21 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.aws.greengrass.mqtt.bridge.clients;
+
+/**
+ * Exception thrown by the MQTT Client.
+ */
+public class MQTTClientException extends MessageClientException {
+    static final long serialVersionUID = -3387516993124229948L;
+
+    MQTTClientException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+    MQTTClientException(String msg) {
+        super(msg);
+    }
+}
