+++ src/test/resources/com/aws/greengrass/mqtt/bridge/config.yaml
