+++ src/test/resources/com/aws/greengrass/mqtt/bridge/config_with_mapping.yaml
